third_party/blink/web_tests/fast/forms/range/range-drag-when-toggled-disabled-expected.txt

Test for dragging operations of <input type=range> when readonly or disabled are toggled.

Tests for range dragging while it toggles to be readonly.
readOnly=false, disabled=false
PASS input.value is "100"
PASS lastChangeEventCounter is changeEventCounter
PASS input.value is "0"
PASS changeEventCounter is >= lastChangeEventCounter + 1
readOnly=true
PASS input.value is "0"
PASS lastChangeEventCounter is changeEventCounter


Tests for range dragging while it toggles to be disabled.
readOnly=false, disabled=false
PASS input.value is "100"
PASS lastChangeEventCounter is changeEventCounter
PASS input.value is "0"
PASS changeEventCounter is >= lastChangeEventCounter + 1
disabled=true
PASS input.value is "0"
PASS lastChangeEventCounter is changeEventCounter


PASS successfullyParsed is true

TEST COMPLETE

 
